Add BFQ-v8r12
This commit is the result of the following operations.

1. The squash of all the commits between "block: cgroups, kconfig,
build bits for BFQ-v7r11-4.5.0" and BFQ-v8r12 in the branch
bfq-mq-v8-v4.11

2. The renaming of two files (block/bfq-cgroup.c ->
block/bfq-cgroup-included.c and block/bfq-iosched.c ->
block/bfq-sq-iosched.c) and of one option (CONFIG_BFQ_GROUP_IOSCHED ->
CONFIG_BFQ_SQ_GROUP_IOSCHED), to avoid name clashes. These name
clashes are due to the presence of bfq in mainline from 4.12.

3. The modification of block/Makefile and block/Kconfig.iosched to
comply with the above renaming.

4. The integration, in the documentation, of information about which
kernel version this code is for, and how this code has been
possibly obtained from code for previous kernel versions. This
information may be very useful for backporting to older kernels.

Signed-off-by: Mauro Andreolini <mauro.andreolini@unimore.it>
Signed-off-by: Arianna Avanzini <avanzini@google.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: Paolo Valente <paolo.valente@linaro.org>

Signed-off-by: Carlos Jimenez (JavaShin-X) <javashin1986@gmail.com>